+++ README.md
@@ -1,36 +1,40 @@
-This is a [Next.js](https://nextjs.org) project bootstrapped with [`create-next-app`](https://nextjs.org/docs/app/api-reference/cli/create-next-app).
+# Git
 
-## Getting Started
+É a ferramenta de controle de versão;
+Ela é a linha do tempo do Software;
 
-First, run the development server:
+# Github
 
-```bash
-npm run dev
-# or
-yarn dev
-# or
-pnpm dev
-# or
-bun dev
-```
+É a plataforma de hospedagem de repositórios Git;
 
-Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
+---
 
-You can start editing the page by modifying `app/page.js`. The page auto-updates as you edit the file.
+JSX
 
-This project uses [`next/font`](https://nextjs.org/docs/app/building-your-application/optimizing/fonts) to automatically optimize and load [Geist](https://vercel.com/font), a new font family for Vercel.
+# HTML e CSS
 
-## Learn More
+São linguagens de marcação e estilo;
 
-To learn more about Next.js, take a look at the following resources:
+# JavaScript
 
-- [Next.js Documentation](https://nextjs.org/docs) - learn about Next.js features and API.
-- [Learn Next.js](https://nextjs.org/learn) - an interactive Next.js tutorial.
+É a linguagem de programação;
 
-You can check out [the Next.js GitHub repository](https://github.com/vercel/next.js) - your feedback and contributions are welcome!
+---
 
-## Deploy on Vercel
+# React
 
-The easiest way to deploy your Next.js app is to use the [Vercel Platform](https://vercel.com/new?utm_medium=default-template&filter=next.js&utm_source=create-next-app&utm_campaign=create-next-app-readme) from the creators of Next.js.
+React é uma biblioteca JavaScript para criar interfaces de usuário;
 
-Check out our [Next.js deployment documentation](https://nextjs.org/docs/app/building-your-application/deploying) for more details.
+# Next
+
+Framework React para criar aplicações web;
+
+# NodeJS
+
+Onde você roda o JavaScript fora do navegador;
+
+---
+
+# Wrapper
+
+Uma ferramente que simplesmente adiciona funcionalidades a outra ferramenta;
