+++ src/components/square.js
@@ -1,0 +1,10 @@
+export function Square({ value, onSquareClick, isWinning }) {
+  return (
+    <button
+      className={`square ${isWinning ? "winning" : ""}`}
+      onClick={onSquareClick}
+    >
+      {value}
+    </button>
+  );
+}
